# docs/open-api-spec/swagger.yaml
definitions:
...
  controllers.DLQList:
    properties:
      deadJobs:
        items:
          $ref: '#/definitions/controllers.DeadDeliveryJobModel'
        type: array
      pages:
        additionalProperties:
          type: string
        type: object
    type: object
...
  controllers.DeadDeliveryJobModel:
    properties:
      jobRequeueURL:
        type: string
      listenerEndpoint:
        type: string
      listenerName:
        type: string
      messageURL:
        type: string
      status:
        type: string
      statusChangedAt:
        type: string
    type: object
  controllers.DeliveryJobModel:
    properties:
      listenerEndpoint:
        type: string
      listenerName:
        type: string
      status:
        type: string
      statusChangedAt:
        type: string
    type: object
...
  controllers.ListResult:
    properties:
      links:
        additionalProperties:
          type: string
        type: object
      pages:
        additionalProperties:
          type: string
        type: object
      result:
        items:
          type: string
        type: array
    type: object
  controllers.MessageModel:
    properties:
      contentType:
        type: string
      dispatchedAt:
        type: string
      headers:
        $ref: '#/definitions/data.HeadersMap'
      id:
        type: string
      jobs:
        items:
          $ref: '#/definitions/controllers.DeliveryJobModel'
        type: array
      payload:
        type: string
      producedBy:
        type: string
      receivedAt:
        type: string
      status:
        type: string
    type: object
  controllers.MessagesStatusResponse:
    properties:
      counts:
        additionalProperties:
          $ref: '#/definitions/controllers.TheCount'
        type: object
      next_scheduled_message_at:
        type: string
    type: object
...
  controllers.TheCount:
    properties:
      count:
        type: integer
      links:
        additionalProperties:
          type: string
        type: object
    type: object
...
  data.HeadersMap:
    additionalProperties:
      type: string
    type: object
info:
  contact: {}
  description: This is a fully HTTP-based Pub/Sub Broker designed to simplify system
    architecture in SOA or Microservice architecture.
  title: Webhook Broker API
  version: 1.0.0
paths:
...
  /channel/{channelId}/consumer/{consumerId}/dlq:
    delete:
      description: Purge all dead jobs for a consumer.
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Consumer ID
        in: path
        name: consumerId
        required: true
        type: string
      - description: Channel authentication token
        in: header
        name: X-Broker-Channel-Token
        required: true
        type: string
      - description: Consumer authentication token
        in: header
        name: X-Broker-Consumer-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deletedCount
          schema:
            type: object
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - ChannelToken: []
        ConsumerToken: []
      summary: Purge Dead Jobs
      tags:
      - DLQ
    get:
      description: Retrieves all dead (failed) delivery jobs for a consumer.
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Consumer ID
        in: path
        name: consumerId
        required: true
        type: string
      - description: Pagination cursor for previous page
        in: query
        name: previous
        type: string
      - description: Pagination cursor for next page
        in: query
        name: next
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.DLQList'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get Dead Jobs for Consumer
      tags:
      - DLQ
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Requeues all dead jobs for a consumer for another delivery attempt.
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Consumer ID
        in: path
        name: consumerId
        required: true
        type: string
      - description: Consumer token for requeue validation
        in: formData
        name: requeue
        required: true
        type: string
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "415":
          description: Unsupported Media Type
        "500":
          description: Internal Server Error
      summary: Requeue All Dead Jobs
      tags:
      - DLQ
...
  /channel/{channelId}/message/{messageId}:
    get:
      description: Retrieves details of a specific message including all delivery
        jobs.
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Message ID
        in: path
        name: messageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MessageModel'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get Message Details
      tags:
      - Messages
  /channel/{channelId}/messages:
    get:
      description: Retrieves a paginated list of messages for a channel with optional
        status filtering.
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Filter by message status
        in: query
        name: status
        type: integer
      - description: Pagination cursor for previous page
        in: query
        name: previous
        type: string
      - description: Pagination cursor for next page
        in: query
        name: next
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ListResult'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: List Messages for a Channel
      tags:
      - Messages
  /channel/{channelId}/messages-status:
    get:
      description: Retrieves message status counts for a channel, including both regular
        and scheduled messages.
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MessagesStatusResponse'
        "500":
          description: Internal Server Error
      summary: Get Message Status Counts
      tags:
      - Messages
...
securityDefinitions:
  ChannelToken:
    description: Channel authentication token issued when the channel was created
      or last rotated.
    in: header
    name: X-Broker-Channel-Token
    type: apiKey
  ConsumerToken:
    description: Consumer authentication token issued when the consumer was created
      or last rotated.
    in: header
    name: X-Broker-Consumer-Token
    type: apiKey
  ProducerID:
    description: Identifier of the producer originating the request.
    in: header
    name: X-Broker-Producer-ID
    type: apiKey
  ProducerToken:
    description: Producer authentication token issued when the producer was created
      or last rotated.
    in: header
    name: X-Broker-Producer-Token
    type: apiKey
swagger: "2.0"